Allow to control showing grid rows and columns
The grid columns can be distracting when debugging the layout using
other tools, like for instance the svg grid from Susy:

  #page {
    @include set-container;
    background: susy-svg-grid() no-repeat scroll;
  }

So add two new configuration variables ($debug-grid-show-rows and
$debug-grid-show-columns) to control showing rows and columns of the
megatype debug grid.

The default behavior is unchanged, the user has to explicitly disable
the display of rows and columns in their configuration.